added users model and linked it to tasks model, added jwt token authentication

+++ app.py
@@ -1,47 +1,78 @@
 from flask import Flask, request, jsonify
 from flask_sqlalchemy import SQLAlchemy
 import json
+import os
+from datetime import timedelta
+from flask_jwt_extended import (
+    JWTManager,
+    create_access_token,
+    jwt_required,
+    get_jwt_identity,
+    create_refresh_token
+)
+from dotenv import load_dotenv
 
+load_dotenv()
 app = Flask(__name__)
+
+#######################
+##TOKEN CONFIGRATIONS##
+#######################
+
+app.config['JWT_SECRET_KEY'] = os.environ.get('JWT_SECRET_KEY')
+app.config['JWT_ACCESS_TOKEN_EXPIRES'] = timedelta(days=1)
+
+jwt = JWTManager(app)
 
 #######################
 ###DB  CONFIGRATIONS###
 #######################
 
-#change pw for security
-DATABASE_URI = 'postgresql://postgres:password@localhost:5432/flask_todo'
-app.config['SQLALCHEMY_DATABASE_URI'] = DATABASE_URI
+app.config['SQLALCHEMY_DATABASE_URI'] = os.environ.get('DATABASE_URI')
 db = SQLAlchemy(app)
 
 #######################
 ####### MODELS ########
 #######################
 
+class User(db.Model):
+    __tablename__ = 'users'
+    id = db.Column(db.Integer, primary_key=True, autoincrement=True)
+    username = db.Column(db.String)
+    password = db.Column(db.String)
+
+    def __repr__(self):
+        return self.username
 class Task(db.Model):
-    __table_name__ = 'task'
+    __tablename__ = 'task'
     id = db.Column(db.Integer, primary_key=True, autoincrement=True)
     title = db.Column(db.String)
     description = db.Column(db.String)
     status = db.Column(db.String, nullable=False)
+    user_id = db.Column(db.Integer, db.ForeignKey('users.id'))
+    user = db.relationship("User", foreign_keys=[user_id])
 
+    def __repr__(self):
+        return self.title + ' - ' + self.status
 #######################
 ## HELPER FUNCTIONS ###
 #######################
 
-def serialize_tasks(queryset):
+def serializer_list(queryset, attrs):
     serialized_list = []
     for obj in queryset:
-        task_dict = serialize_task(obj)
+        task_dict = serializer(obj, attrs)
         serialized_list.append(task_dict)
     return serialized_list
 
-def serialize_task(obj):
-    task_dict = {}
-    task_dict['id'] = obj.id
-    task_dict['title'] = obj.title
-    task_dict['description'] = obj.description
-    task_dict['status'] = obj.status
-    return task_dict
+
+#attrs is list of strings representing attributes in object
+def serializer(obj , attrs):
+    serialized_obj = {}
+    for attr in attrs:
+        serialized_obj[attr] = getattr(obj,attr)
+    return serialized_obj
+
 
 def update_task_with_validation(task,data):
     keys = ['title','description','status']
@@ -55,38 +86,107 @@ def update_task_with_validation(task,data):
 ######## APIS #########
 #######################
 
+@app.route('/register', methods=['POST'])
+def register():
+    try:
+        data = json.loads(request.data)
+        user = User(username=data['username'],password=data['password'])
+        db.session.add(user)
+        db.session.commit()
+        return jsonify({
+        'status': 'Success',
+        'msg': "user created successfully",
+        'data': serializer(user,['id','username','password'])
+        }), 201
+    except:
+        return jsonify({
+        'status': 'Fail',
+        'msg': "Incorrect user data"
+        }), 400
+
+
+@app.route('/login', methods=['POST'])
+def login():
+    username = request.json.get('username')
+    password = request.json.get('password')
+    try:
+        db_user = User.query.filter_by(username=username).first()
+        if db_user.password == password:
+            access_token = create_access_token(identity=db_user.id)
+            refresh_token = create_refresh_token(identity=db_user.id)
+
+            return jsonify({
+                'status': 'success',
+                'data': {
+                    'access_token': access_token,
+                    'refresh_token': refresh_token
+                }
+            }), 200
+        return jsonify({
+                'status': 'Fail',
+                'msg': "username or password is incorrect"
+            }), 403
+    except:
+        return jsonify({
+                'status': 'Fail',
+                'msg': "No such user with the given username"
+            }), 404
+
 @app.route('/', methods=['GET'])
 def tasks():
     tasks_qs = Task.query.all()
-    tasks = serialize_tasks(tasks_qs)
+    tasks = serializer_list(tasks_qs,['id','title','description','status','user_id'])
     return jsonify({
         "data":tasks
     })
 
 @app.route('/create_task', methods=['POST'])
+@jwt_required()
 def create_task():
     data = json.loads(request.data)
-    task = Task(title=data['title'], description=data['description'], status=data['status'])
-   
-    db.session.add(task)
-    db.session.commit()
+    user_id = get_jwt_identity()
+    if User.query.filter_by(id=user_id).first(): 
+        try:
+            task = Task(title=data['title'], description=data['description'],
+            status=data['status'],user_id=user_id)
+        
+            db.session.add(task)
+            db.session.commit()
 
-    return jsonify({
-        "data":data,
-        "msg":"Task Created Successfully"
-    }), 201
+            serialized_task = serializer(obj=task,attrs=['id','title',
+            'description','status','user_id'])
+            return jsonify({
+                "data": serialized_task,
+                "msg":"Task Created Successfully"
+            }), 201
+        except:
+            return jsonify({
+            "msg":"Invalid task data was recieved"
+        }), 400
+    else:
+        return jsonify({
+            "msg":"wrong user id"
+        }), 400
+
 
 @app.route('/update_task/<int:id>', methods=['PUT'])
+@jwt_required()
 def update_task(id):
 
-    data = json.loads(request.data)
     task = Task.query.filter_by(id=id).first()
-
     if not task:
         return jsonify({
             "msg":"Didn't find any tasks with given id"
         })
 
+    data = json.loads(request.data)
+    user_id = get_jwt_identity()
+
+    if task.user_id != user_id:
+        return jsonify({
+                "msg":"Your are unauthorized to update this task"
+            }), 403
+    
     if update_task_with_validation(task,data) is not None:
         return jsonify({
                 "msg":"There was an error in the sent data"
@@ -95,25 +195,42 @@ def update_task(id):
     db.session.commit()
 
     return jsonify({
-        "data":serialize_task(task),
+        "data":serializer(task,['id','title','description','status','user_id']),
         "msg":"Updated Task Successfully"
     }), 200
 
 @app.route('/delete_task/<int:id>', methods=['DELETE'])
+@jwt_required()
 def delete_task(id):
 
     task = Task.query.filter_by(id=id).first()
-
     if not task:
         return jsonify({
             "msg":"Didn't find any tasks with given id"
         })
 
+    user_id = get_jwt_identity()
+
+    if task.user_id != user_id:
+        return jsonify({
+                "msg":"Your are unauthorized to delete this task"
+            }), 403
+            
     db.session.delete(task)
     db.session.commit()
 
     return jsonify({
         "msg":"Deleted Task Successfully"
+    }), 200
+
+@app.route('/refresh', methods=['POST'])
+@jwt_required(refresh=True)
+def refresh():
+    user_id = get_jwt_identity()
+    access_token = create_access_token(identity=user_id)
+
+    return jsonify({
+        'access_token': access_token
     }), 200
 
 db.create_all()
